fix presenter and ModelSelectd
