+++ src/Main.java
@@ -1,4 +1,4 @@
-import co.edu.uptc.presenters.ManagerGeneral;
+import co.edu.uptc.presenter.ManagerGeneral;
 
 public class Main {
     public static void main(String[] args) {
