+++ src/co/edu/uptc/models/Model202127343/ManagerModel202127343.java
@@ -1,7 +1,7 @@
 package co.edu.uptc.models.Model202127343;
 
 import co.edu.uptc.pojos.Person;
-import co.edu.uptc.presenters.ContratBills;
+import co.edu.uptc.presenter.ContratBills;
 
 import java.util.List;
