+++ src/co/edu/uptc/models/Model_202128710/Manager_Model_202128710.java
@@ -2,7 +2,7 @@ package co.edu.uptc.models.Model_202128710;
 
 
 import co.edu.uptc.pojos.Person;
-import co.edu.uptc.presenters.ContratBills;
+import co.edu.uptc.presenter.ContratBills;
 
 import java.util.List;
