+++ src/co/edu/uptc/models/model202127061/ManagerModel202127061.java
@@ -1,7 +1,7 @@
 package co.edu.uptc.models.model202127061;
 
 import co.edu.uptc.pojos.Person;
-import co.edu.uptc.presenters.ContratBills;
+import co.edu.uptc.presenter.ContratBills;
 
 import java.util.List;
