+++ src/co/edu/uptc/models/model_202115100/ModelSantiago_202115100.java
@@ -1,7 +1,7 @@
 package co.edu.uptc.models.model_202115100;
 
 import co.edu.uptc.pojos.Person;
-import co.edu.uptc.presenters.ContratBills;
+import co.edu.uptc.presenter.ContratBills;
 
 import java.util.List;
